# backend/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "NeutriAI API",
    "description": "AI nutrition, fitness and lifestyle platform. All endpoints except `/healthz`, `/v1/billing/plans` and `/v1/webhooks/stripe` require a Supabase bearer token.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/v1/scans": {
      "post": {
        "tags": [
          "nutrition"
        ],
        "summary": "Create Scan",
        "description": "Analyse an uploaded meal photo.\n\nThe client uploads to the ``meal-photos`` bucket under ``{user_id}/...``\nfirst, then posts the object keys here. Runs inline because the user is\nstaring at a spinner; typical latency is 4-8 s.\n\nTHE PREFIX IS CHECKED HERE, and this docstring used to claim RLS did it.\nRLS governs the UPLOAD. The read further down uses the service key, which\nbypasses both RLS and the storage policy by design, so nothing between this\nline and the download ever compared the path to the caller.\n\n    POST /v1/scans {\"image_paths\": [\"<someone-else-uuid>/IMG_0042.jpg\"]}\n\nreturned that person's meal, analysed, to whoever asked. Their photo, their\nfood, their plate. One prefix comparison closes it.",
        "operationId": "create_scan_v1_scans_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScanRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScanResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/scans/{scan_id}": {
      "get": {
        "tags": [
          "nutrition"
        ],
        "summary": "Get Scan",
        "operationId": "get_scan_v1_scans__scan_id__get",
        "parameters": [
          {
            "name": "scan_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Scan Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/calibrations": {
      "get": {
        "tags": [
          "nutrition"
        ],
        "summary": "List Calibrations",
        "operationId": "list_calibrations_v1_calibrations_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "nutrition"
        ],
        "summary": "Add Calibration",
        "operationId": "add_calibration_v1_calibrations_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CalibrationIn"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/calibrations/progress": {
      "get": {
        "tags": [
          "nutrition"
        ],
        "summary": "Calibration Progress",
        "description": "How well this user's own crockery is measured, and what would help next.\n\nThe headline is the ACCURACY, not a count of photos. \"12 of 20 photos\" is a\nmeasure of the user's compliance; \"your dinner plate is measured to 2.4%\"\nis a measure of what they are getting for it, and it is the number that\ndecides whether their portions are right. Scale multiplies every gram: a\nwidth 10% wrong makes every portion off that plate 21% wrong in area before\nthe food has even been named.\n\nEverything here is computed from the stored observations at request time,\nnot from a counter someone incremented. A number that cannot be recomputed\ncannot be trusted, and this one is shown to the user as a promise.",
        "operationId": "calibration_progress_v1_calibrations_progress_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/calibrations/measure": {
      "post": {
        "tags": [
          "nutrition"
        ],
        "summary": "Record Measurement",
        "description": "The user measured a vessel with a tape. Worth about 28 photos.\n\nKept separate from POST /calibrations, which writes a size directly. This\nfiles an OBSERVATION, so a tape reading sits in the same history as every\ninferred one and the published width and its error are recomputed from the\nwhole record. It also means a mistyped measurement can be seen and removed\nrather than silently becoming the answer.",
        "operationId": "record_measurement_v1_calibrations_measure_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VesselMeasurementIn"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/meals": {
      "get": {
        "tags": [
          "nutrition"
        ],
        "summary": "List Meals",
        "operationId": "list_meals_v1_meals_get",
        "parameters": [
          {
            "name": "day",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Day"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MealOut"
                  },
                  "title": "Response List Meals V1 Meals Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "nutrition"
        ],
        "summary": "Create Meal",
        "description": "Manual meal entry, or logging a saved recipe.",
        "operationId": "create_meal_v1_meals_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MealIn"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MealOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/assessments": {
      "get": {
        "tags": [
          "nutrition"
        ],
        "summary": "List Assessments",
        "operationId": "list_assessments_v1_assessments_get",
        "parameters": [
          {
            "name": "day",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Day"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AssessmentOut"
                  },
                  "title": "Response List Assessments V1 Assessments Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/foods/search": {
      "get": {
        "tags": [
          "nutrition"
        ],
        "summary": "Search Foods",
        "description": "Type-ahead over the shared food cache, falling back to a live lookup.",
        "operationId": "search_foods_v1_foods_search_get",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 2,
              "title": "Q"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "VesselMeasurementIn": {
        "properties": {
          "vessel": {
            "type": "string",
            "maxLength": 40,
            "minLength": 1,
            "title": "Vessel"
          },
          "width_mm": {
            "type": "number",
            "maximum": 600.0,
            "minimum": 60.0,
            "title": "Width Mm"
          }
        },
        "type": "object",
        "required": [
          "vessel",
          "width_mm"
        ],
        "title": "VesselMeasurementIn",
        "description": "One vessel measured with a tape.\n\nBounds are physical, not defensive: a dinner plate is not 4 cm across and\nnot 60 cm across, and a typo either side of that would otherwise become the\nscale for every meal the user eats off it."
      },
...
    }
  }
}